Add solution without using recursion
I found this the easiest.
I know how `while` loops work and I have faith that this one will get
to every single sub-directory.
Scanning through one directory before moving to the next is also
something that is easy to wrap my head around.

I have to keep appending to a list though. That seems like a downside
compared to an approach that uses generators.